working null gravity
